requirements.txt
pynput
opencv-python
pywin32
pillow
numpy==1.26.4
setuptools
pyyaml
pydantic
onnxruntime-gpu
paddlepaddle-gpu
paddleocr==2.7.3
colorama
psutil
future
requests
pydantic
gitpython
pystray
PyQt5
screeninfo